refactor: delete legacy store files and add migration to drop tables
