+++ assistant/src/__tests__/guardian-outbound-http.test.ts
@@ -157,7 +157,6 @@ afterAll(() => {
 
 function resetTables(): void {
   const db = getDb();
-  db.run("DELETE FROM channel_guardian_bindings");
   db.run("DELETE FROM channel_guardian_verification_challenges");
   try {
     db.run("DELETE FROM channel_guardian_approval_requests");
