+++ assistant/src/__tests__/inbound-invite-redemption.test.ts
@@ -113,12 +113,10 @@ let msgCounter = 0;
 
 function resetState(): void {
   const db = getDb();
-  db.run("DELETE FROM assistant_ingress_members");
   db.run("DELETE FROM assistant_ingress_invites");
   db.run("DELETE FROM channel_inbound_events");
   db.run("DELETE FROM conversations");
   db.run("DELETE FROM channel_guardian_approval_requests");
-  db.run("DELETE FROM channel_guardian_bindings");
   db.run("DELETE FROM notification_events");
   db.run("DELETE FROM contact_channels");
   db.run("DELETE FROM contacts");
