+++ assistant/src/__tests__/ingress-routes-http.test.ts
@@ -48,7 +48,6 @@ afterAll(() => {
 });
 
 function resetTables() {
-  getSqlite().run("DELETE FROM assistant_ingress_members");
   getSqlite().run("DELETE FROM assistant_ingress_invites");
   getSqlite().run("DELETE FROM contact_channels");
   getSqlite().run("DELETE FROM contacts");
