+++ assistant/src/__tests__/invite-redemption-service.test.ts
@@ -47,7 +47,6 @@ afterAll(() => {
 });
 
 function resetTables() {
-  getSqlite().run("DELETE FROM assistant_ingress_members");
   getSqlite().run("DELETE FROM assistant_ingress_invites");
   getSqlite().run("DELETE FROM contact_channels");
   getSqlite().run("DELETE FROM contacts");
