+++ assistant/src/__tests__/trusted-contact-lifecycle-notifications.test.ts
@@ -120,13 +120,11 @@ const GUARDIAN_APPROVAL_TTL_MS = 5 * 60 * 1000;
 function resetState(): void {
   const db = getDb();
   db.run("DELETE FROM channel_guardian_approval_requests");
-  db.run("DELETE FROM channel_guardian_bindings");
   db.run("DELETE FROM channel_guardian_verification_challenges");
   db.run("DELETE FROM channel_guardian_rate_limits");
   db.run("DELETE FROM channel_inbound_events");
   db.run("DELETE FROM conversations");
   db.run("DELETE FROM notification_events");
-  db.run("DELETE FROM assistant_ingress_members");
   db.run("DELETE FROM contact_channels");
   db.run("DELETE FROM contacts");
   emitSignalCalls.length = 0;
