+++ assistant/src/__tests__/trusted-contact-multichannel.test.ts
@@ -128,13 +128,11 @@ const TEST_BEARER_TOKEN = "test-token";
 function resetState(): void {
   const db = getDb();
   db.run("DELETE FROM channel_guardian_approval_requests");
-  db.run("DELETE FROM channel_guardian_bindings");
   db.run("DELETE FROM channel_guardian_verification_challenges");
   db.run("DELETE FROM channel_guardian_rate_limits");
   db.run("DELETE FROM channel_inbound_events");
   db.run("DELETE FROM conversations");
   db.run("DELETE FROM notification_events");
-  db.run("DELETE FROM assistant_ingress_members");
   db.run("DELETE FROM contact_channels");
   db.run("DELETE FROM contacts");
   emitSignalCalls.length = 0;
