+++ assistant/src/__tests__/voice-invite-redemption.test.ts
@@ -42,7 +42,6 @@ afterAll(() => {
 });
 
 function resetTables() {
-  getSqlite().run("DELETE FROM assistant_ingress_members");
   getSqlite().run("DELETE FROM assistant_ingress_invites");
   getSqlite().run("DELETE FROM contact_channels");
   getSqlite().run("DELETE FROM contacts");
