+++ assistant/src/runtime/actor-trust-resolver.ts
@@ -18,7 +18,7 @@ import {
 } from "../contacts/contact-store.js";
 import { contactChannelToMemberRecord } from "../contacts/member-record-shim.js";
 import type { TrustContext } from "../daemon/session-runtime-assembly.js";
-import type { IngressMember } from "../memory/ingress-member-store.js";
+import type { IngressMember } from "../contacts/member-record-shim.js";
 import { canonicalizeInboundIdentity } from "../util/canonicalize-identity.js";
 import { getLogger } from "../util/logger.js";
